Update: remove unused files & select tag func
