+++ src/components/post/Tag.tsx
@@ -1,23 +1,25 @@
 import React from "react";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import { darken, lighten } from "polished";
 
 type TagProps = {
   // key: string;
   title: string;
   onClick: () => void;
+  selected: boolean;
 }
 
 // const Tag = ({ title, onClick }: TagProps) => <TagContainer onClick={onClick}>{title}</TagContainer>;
-const Tag = ({ title, onClick }: TagProps) => {
-  return <TagContainer onClick={onClick}>{title}</TagContainer>;
+const Tag = ({ title, onClick, selected }: TagProps) => {
+  console.log(selected);
+  return <TagContainer onClick={onClick} selected={selected}>{title}</TagContainer>;
 };
 
 export default React.memo(Tag);
 // export default Tag;
 
 
-const TagContainer = styled.li`
+const TagContainer = styled.li<Pick<TagProps, "selected">>`
   cursor: pointer;
   font-weight: 700;
   font-size: 1.2rem;
@@ -27,9 +29,10 @@ const TagContainer = styled.li`
   /* color: var(--white); */
   /* background-color: var(--purple); */
 
-  .active {
-    background: var(--teal);
-  }
+  ${({ selected }) => selected && css`
+    background-color: var(--teal);
+  `}
+
 
   &:hover {
     color: ${({ theme }) => theme ? lighten(0.1, theme.themes.color.purple) : "var(--white)"};
