--- src/layouts/Footer.tsx
@@ -1,10 +1,0 @@
-const Footer = () => {
-
-  return (
-    <footer>
-      Hello Footer??
-    </footer>
-  )
-}
-
-export default Footer;
